Make it configurable if the dma map call is made only once
If we map on every request it may happen that we run out on kernel
memory for the mappings. The reason is that the kernel does not clear
intermediate page table entries and the current algorithm for kdma
mappings not always is able to reuse virtual addresses.

Fix this by mapping every ds only once. This gives also some performance
improvements. Please note that this makes the full ds memory available
to the driver and the device.

Change-Id: I99157a28c11c822a5b17a88c804006249eb85ea4