properly initialize Emmc::Drv::_bb_size
This variable could be uninitialized if no bounce buffer capability was
passed to the driver. Initialize this variable and introduce a dedicated
function `using_bounce_buffer()` to test if we are using a bounce buffer
instead of testing `_bb_size` for 0.

Also fix the logging in `Sdhci::adma2_set_desc()`.

See ticket #CD-930.

Change-Id: I993d3b315ff51709ef059ebe1eab8b0e480663ee