drv_sdhci: overhaul delayed writing
Be more accurate in delayed writing and remember the last time we wrote
to some register. Limit the amount of waiting time to some microseconds.
In fact it is sufficient to wait no more than 3-4 SD card cycles.

Unfortunately we cannot store any information related to delayed writing
in a static variable because then different controllers driven by the
same driver application could interfere. Instead, store the information
about the delay and the last access time in the driver instance and pass
the 'this' pointer to Sdhci::Reg::write(), and, for consistency, also to
Sdhci::Reg::read().

The delayed writing seems to be required the whole time, not only during
driver initialization. The overhead should be acceptable because usually
the frequency of register write accesses to the device is not that high.

Change-Id: I6ca9f8d39af81681d9611ab76e82f15768eef784