don't pretend to use bounce buffer if not supported by the driver
Warn if a bounce buffer is provided ("bbds" capability) but the driver
doesn't make use of it.

A bounce buffer is only used by the SDHCI driver for requests where the
physical buffer resides beyond 4GB.

Also print the maximum request size because this value is limited by the
bounce buffer.

Change-Id: Iacc74af5449a1917776b9e70ab558abae9fa7029