improve SDIO interface handling
Always reset an SDIO card to show only its SD memory interface before
sending the first CMD0 to the device. Also do not send CMD5 to the
device because this command would activate the IO interface of an SDIO
card.

So far the SDIO reset was only performed for iproc/arasan SDHCI devices.

Change-Id: I763074944125a2a5410f9afd34a13df5bdae9983