sdhci: consider DMA limit
Consider the DMA limit if applicable. For client memory we can use the
bounce buffer. But the bounce buffer, the IO buffer (CMD8 etc) and the
DMA descriptor buffer have to consider the DMA limit as well. The DMA
memory for the BCM2835 device has to be located below 1GB.

Check the physical address of these buffers and panic if the DMA limit
is not fulfilled. Memory for these buffers can be provided via external
dataspace capabilities.

Change-Id: I9ebb40bb27a7712fe82fe95bab7d1455131a2975