also consider the size when mapping MMIO resources
Consider the size of the MMIO region. In most cases, assuming a 4K
region is sufficient but still it's better to use the correct values.

Change the type of MMIO address ranges to 64-bit to be on the safe side.
Also eager-map the whole MMIO region because these regions are mapped on
the first access anyway.

Change-Id: I578a679d1209c1ea1cb50ab036361b5d0b9d3505