support tuning for UHS SDR104, UHS SDR50 and UHS DDR50 in SD card mode
Tuning is required for high-speed UHS timings for SD cards as well,
similar to tuning support for eMMC devices.

Tuning is always required for UHS SDR104 timing. Tuning is required for
UHS SDR50 timing if this requirement is announced by the controller. We
assume that tuning is always required for UHS DDR50 as well although
tuning requirements are not documented for this mode. Linux seems to
perform tuning for all three timings unconditionally.

Also enable auto-tuning for uSDHC controllers after successful tuning.

See ticket #CD-1402.

Change-Id: I107882982c05a9f992c6ba136390f09423032e2f