Provide also the offset to Dma_space::map
This fixes the case when the l4virtio driver memory region is registered
with non-zero dataspace offset. With the possibly non-zero ds_offset(),
it is also necessary to pass the actual size of the memory region and
not the size of the entire dataspace. As a bonus, this saves one RPC on
the first dma_map_all operation.

Change-Id: I3b0edd6364b95124e54d20ab6ec98bb33920c853