make '--dma-map-all' the default behavior
So far the default behavior was to map the relevant part of a client
dataspace before an I/O request and to unmap it after the request.

Change the default behavior to map the entire client dataspace into the
DMA space at the first I/O request and never unmap is until the client
is destroyed.

The switch '--dma-map-all' is still supported.
The switch '--dma-map-per-req' restores the old behavior.

Change-Id: I03791f4fb125b8ba9893c9ccbd3b1c69f2fd9ef3