drv_sdhci: max_seg defines size of ADMA2 memory and limits max_size()
The maximum inout request size is limited by the device. The maximum
number of segments per request and the maximum request size determine
the amount of memory which is required as memory for AMDA2 descriptors.

Replace the previously fixed number of 4096 for the ADMA2 descriptor
memory with the correct one.

Change-Id: I6420ce855a07a0ebd2bf046c0460d3e4a1b3223d
Signed-off-by: Frank Mehnert <frank.mehnert@kernkonzept.com>